Refactor email sending component
It was too big and not easily reusable...
